requirements.txt
fastapi
uvicorn
sqlalchemy
 mysql-connector-python
paho-mqtt
cryptography
python-dotenv
